chapter14/requirements.txt
aniso8601==8.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-RESTful==0.3.8
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
luma.core==1.12.0
luma.led-matrix==1.4.1
MarkupSafe==1.1.1
paho-mqtt==1.5.0
pigpio==1.45
Pillow==7.1.0
pyftdi==0.44.0
Pypubsub==4.0.3
pyserial==3.4
python-engineio==3.11.2
python-socketio==4.4.0
pytz==2019.3
pyusb==1.0.2
requests==2.22.0
rpi-ws281x==4.2.3
RPi.GPIO==0.7.0
six==1.14.0
smbus2==0.3.0
spidev==3.4
urllib3==1.25.8
Werkzeug==1.0.0
ws2812==1.0.0
